@@ -11,8 +11,8 @@ allprojects {
     group = 'com.apriori.qa'
     version = '0.0.1-SNAPSHOT'
     apply plugin: 'java'
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
 }
 
 subprojects {
@@ -129,7 +129,7 @@ subprojects {
         implementation 'org.junit.jupiter:junit-jupiter:5.9.3'
         implementation 'org.junit.platform:junit-platform-suite-engine:1.9.2'
         testImplementation 'org.junit.platform:junit-platform-launcher:1.9.3'
-        implementation 'io.github.bonigarcia:webdrivermanager:5.3.3'
+        implementation 'io.github.bonigarcia:webdrivermanager:5.4.1'
         implementation 'org.seleniumhq.selenium:selenium-java:4.11.0'
         testImplementation 'io.qameta.allure:allure-junit5:2.22.0'
         implementation 'org.projectlombok:lombok:1.18.24'
