@@ -4,6 +4,8 @@ import com.apriori.annotations.Schema;
 import com.apriori.deserializers.DateTimeDeserializer_yyyyMMddTHHmmssSSSZ;
 import com.apriori.enums.DigitalFactoryEnum;
 import com.apriori.models.response.CostingInput;
+import com.apriori.models.response.SecondaryDigitalFactories;
+import com.apriori.models.response.SecondaryProcesses;
 
 import com.fasterxml.jackson.annotation.JsonRootName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
@@ -76,8 +78,8 @@ public class CostingTemplate {
     private List<String> propertiesToReset = null;
     private List<RoutingNodeOptions> routingNodeOptions;
     private CostingInput.ProcessSetupOptions processSetupOptions;
-    private CostingInput.SecondaryDigitalFactories secondaryDigitalFactories;
-    private CostingInput.SecondaryProcesses secondaryProcesses;
+    private SecondaryDigitalFactories secondaryDigitalFactories;
+    private SecondaryProcesses secondaryProcesses;
     private CostingInput.Threads threads;
     private CostingInput.Tolerances tolerances;
 }
