@@ -16,13 +16,13 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import utils.Constants;
 
 import java.util.List;
 
 public class ReportsLoginPage extends ReportsPageHeader {
 
     private static final Logger logger = LoggerFactory.getLogger(ReportsLoginPage.class);
-    private static final String loginPageURL = PropertiesContext.get("${env}.reports.ui_url");
     private static final Boolean isEnvOnPrem = PropertiesContext.get("${env}").equals("onprem");
 
     @FindBy(css = "input[name='j_username']")
@@ -77,7 +77,7 @@ public class ReportsLoginPage extends ReportsPageHeader {
 
     @Override
     protected void isLoaded() throws Error {
-        assertThat("CIR login page was not displayed", loginService.getLoginTitle().contains(PropertiesContext.get("${env}.reports.welcome_page_text")));
+        assertThat("CIR login page was not displayed", loginService.getLoginTitle().contains(Constants.WELCOME_PAGE_TEXT));
     }
 
     /**
