@@ -15,8 +15,8 @@ import utils.JasperApiAuthenticationUtil;
 import java.util.Arrays;
 
 public class CostOutlierIdentificationTests extends JasperApiAuthenticationUtil {
-    private static final String reportsJsonFileName = JasperApiEnum.COST_OUTLIER_IDENTIFICATION.getEndpoint();
     private static final String exportSetName = ExportSetEnum.COST_OUTLIER_THRESHOLD_ROLLUP.getExportSetName();
+    private static final String reportsJsonFileName = JasperApiEnum.COST_OUTLIER_IDENTIFICATION.getEndpoint();
     private static final CirApiEnum reportsNameForInputControls = CirApiEnum.COST_OUTLIER_IDENTIFICATION;
     private static JasperApiUtils jasperApiUtils;
 
@@ -26,7 +26,7 @@ public class CostOutlierIdentificationTests extends JasperApiAuthenticationUtil 
     }
 
     @Test
-    @TestRail(id = {26909})
+    @TestRail(id = 26909)
     @Description("Cost metric options available & selected cost metric used in report generated (incl. report header)")
     public void testCostMetricFbcFunctionality() {
         jasperApiUtils.genericCostMetricCostOutlierTest(
@@ -35,7 +35,7 @@ public class CostOutlierIdentificationTests extends JasperApiAuthenticationUtil 
     }
 
     @Test
-    @TestRail(id = {26910})
+    @TestRail(id = 26910)
     @Description("Cost metric options available & selected cost metric used in report generated (incl. report header)")
     public void testCostMetricPpcFunctionality() {
         jasperApiUtils.genericCostMetricCostOutlierTest(
