@@ -12,9 +12,9 @@ import utils.JasperApiAuthenticationUtil;
 
 public class CycleTimeValueTrackingReportTests extends JasperApiAuthenticationUtil {
     private static final String reportsJsonFileName = JasperApiEnum.CYCLE_TIME_VALUE_TRACKING.getEndpoint();
+    private static final CirApiEnum reportsNameForInputControls = CirApiEnum.CYCLE_TIME_VALUE_TRACKING;
     // Export Set is not relevant for this report
     private static final String exportSetName = "";
-    private static final CirApiEnum reportsNameForInputControls = CirApiEnum.CYCLE_TIME_VALUE_TRACKING;
     private static JasperApiUtils jasperApiUtils;
 
     @BeforeEach
@@ -23,7 +23,7 @@ public class CycleTimeValueTrackingReportTests extends JasperApiAuthenticationUt
     }
 
     @Test
-    @TestRail(id = {25987})
+    @TestRail(id = 25987)
     @Description("Verify Currency Code input control is working correctly")
     public void testCurrencyCode() {
         jasperApiUtils.cycleTimeValueTrackingCurrencyTest();
