@@ -1,9 +1,8 @@
 package com.ootbreports.newreportstests.dtcmetrics.sheetmetaldtc;
 
-import static com.apriori.testconfig.TestSuiteType.TestSuite.REPORTS;
-
 import com.apriori.cir.enums.CirApiEnum;
 import com.apriori.enums.ExportSetEnum;
+import com.apriori.enums.ProcessGroupEnum;
 import com.apriori.testrail.TestRail;
 
 import com.ootbreports.newreportstests.utils.JasperApiEnum;
@@ -15,7 +14,6 @@ import enums.MassMetricEnum;
 import enums.SortOrderEnum;
 import io.qameta.allure.Description;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import utils.JasperApiAuthenticationUtil;
 
@@ -26,12 +24,12 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     private static final String reportsJsonFileName = JasperApiEnum.SHEET_METAL_DTC.getEndpoint();
     private static final String exportSetName = ExportSetEnum.SHEET_METAL_DTC.getExportSetName();
     private static final CirApiEnum reportsNameForInputControls = CirApiEnum.SHEET_METAL_DTC;
-    private static JasperApiUtils jasperApiUtils;
     private final List<String> mostCommonPartNames = Arrays.asList(
         JasperCirApiPartsEnum.P_2980123_CLAMP.getPartName(),
         JasperCirApiPartsEnum.AP_BRACKET_HANGER.getPartName(),
         JasperCirApiPartsEnum.DS73_F04604_PIA1.getPartName()
     );
+    private static JasperApiUtils jasperApiUtils;
 
     @BeforeEach
     public void setupGenericMethods() {
@@ -39,7 +37,7 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     }
 
     @Test
-    @TestRail(id = {3046})
+    @TestRail(id = 3046)
     @Description("Verify Currency Code input control functions correctly - Sheet Metal DTC Report")
     public void testCurrencyCode() {
         jasperApiUtils.genericDtcCurrencyTest(
@@ -49,7 +47,7 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     }
 
     @Test
-    @TestRail(id = {3043})
+    @TestRail(id = 3043)
     @Description("Verify cost metric input control functions correctly - PPC - Sheet Metal DTC Report")
     public void testCostMetricInputControlPpc() {
         jasperApiUtils.genericDtcTest(
@@ -59,7 +57,7 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     }
 
     @Test
-    @TestRail(id = {7418})
+    @TestRail(id = 7418)
     @Description("Verify cost metric input control functions correctly - FBC - Sheet Metal DTC Report")
     public void testCostMetricInputControlFbc() {
         jasperApiUtils.genericDtcTest(
@@ -69,7 +67,7 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     }
 
     @Test
-    @TestRail(id = {3044})
+    @TestRail(id = 3044)
     @Description("Verify Mass Metric input control functions correctly - Finish Mass - Sheet Metal DTC Report")
     public void testMassMetricInputControlFinishMass() {
         jasperApiUtils.genericDtcTest(
@@ -79,7 +77,7 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     }
 
     @Test
-    @TestRail(id = {7398})
+    @TestRail(id = 7398)
     @Description("Verify Mass Metric input control functions correctly - Rough Mass - Sheet Metal DTC Report")
     public void testMassMetricInputControlRoughMass() {
         jasperApiUtils.genericDtcTest(
@@ -89,22 +87,21 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     }
 
     @Test
-    @TestRail(id = {7448})
+    @TestRail(id = 7448)
     @Description("Verify process group input control functionality - Single Selection - Sheet Metal DTC Report")
     public void testSingleProcessGroup() {
         List<String> partNames = Arrays.asList(
             JasperCirApiPartsEnum.P_2980123_CLAMP.getPartName(),
             JasperCirApiPartsEnum.AP_BRACKET_HANGER.getPartName()
         );
-        /*jasperApiUtils.genericProcessGroupDtcTest(
+        jasperApiUtils.genericProcessGroupDtcTest(
             partNames,
             "Process Group", ProcessGroupEnum.SHEET_METAL.getProcessGroup()
-        );*/
+        );
     }
 
     @Test
-    @Tag(REPORTS)
-    @TestRail(id = {7532})
+    @TestRail(id = 7532)
     @Description("Verify DTC Score Input Control - Low Selection - Sheet Metal DTC Report")
     public void testDtcScoreLow() {
         List<String> partNames = Arrays.asList(
@@ -120,8 +117,7 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     }
 
     @Test
-    @Tag(REPORTS)
-    @TestRail(id = {7535})
+    @TestRail(id = 7535)
     @Description("Verify DTC Score Input Control - Medium Selection - Sheet Metal DTC Report")
     public void testDtcScoreMedium() {
         List<String> partNames = Arrays.asList(
@@ -137,8 +133,7 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     }
 
     @Test
-    @Tag(REPORTS)
-    @TestRail(id = {7538})
+    @TestRail(id = 7538)
     @Description("Verify DTC Score Input Control - High Selection - Sheet Metal DTC Report")
     public void testDtcScoreHigh() {
         List<String> partNames = Arrays.asList(
@@ -154,8 +149,7 @@ public class SheetMetalDtcReportTests extends JasperApiAuthenticationUtil {
     }
 
     @Test
-    @Tag(REPORTS)
-    @TestRail(id = {3045})
+    @TestRail(id = 3045)
     @Description("Verify Sort Order input control functions correctly - Annual Spend - Sheet Metal DTC Report")
     public void testSortOrderAnnualSpend() {
         List<String> partNames = Arrays.asList(
