@@ -1,13 +1,6 @@
 package com.ootbreports.newreportstests.spendanalysisvaluetracking;
 
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.not;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-import com.apriori.cir.JasperReportSummary;
 import com.apriori.cir.enums.CirApiEnum;
-import com.apriori.enums.CurrencyEnum;
 import com.apriori.testrail.TestRail;
 
 import com.ootbreports.newreportstests.utils.JasperApiEnum;
@@ -17,11 +10,13 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import utils.JasperApiAuthenticationUtil;
 
+import java.util.Arrays;
+
 public class SpendAnalysisValueTrackingDetailsReportTests extends JasperApiAuthenticationUtil {
     private static final String reportsJsonFileName = JasperApiEnum.SPEND_ANALYSIS_VALUE_TRACKING_DETAILS.getEndpoint();
+    private static final CirApiEnum reportsNameForInputControls = CirApiEnum.SPEND_ANALYSIS_VALUE_TRACKING_DETAILS;
     // Export set is irrelevant for this report
     private static final String exportSetName = "";
-    private static final CirApiEnum reportsNameForInputControls = CirApiEnum.SPEND_ANALYSIS_VALUE_TRACKING_DETAILS;
     private static JasperApiUtils jasperApiUtils;
 
     @BeforeEach
@@ -33,23 +28,6 @@ public class SpendAnalysisValueTrackingDetailsReportTests extends JasperApiAuthe
     @TestRail(id = 26932)
     @Description("Input controls - Currency code - Details Report")
     public void testCurrencyCode() {
-        JasperReportSummary gbpJasperReportSummary = jasperApiUtils.genericTestCoreProjectRollupAndCurrencyOnly(
-            CurrencyEnum.GBP.getCurrency(),
-            "AC CYCLE TIME VT 1"
-        );
-
-        String gbpCurrencySettingAboveChart = gbpJasperReportSummary.getReportHtmlPart().getElementsByAttributeValue("colspan", "7").get(2).child(0).text();
-        String gbpInitialApCost = gbpJasperReportSummary.getReportHtmlPart().getElementsByAttributeValue("colspan", "2").get(20).child(0).text();
-
-        JasperReportSummary usdJasperReportSummary = jasperApiUtils.genericTestCoreProjectRollupAndCurrencyOnly(
-            CurrencyEnum.USD.getCurrency(),
-            "AC CYCLE TIME VT 1"
-        );
-
-        String usdCurrencySettingAboveChart = usdJasperReportSummary.getReportHtmlPart().getElementsByAttributeValue("colspan", "7").get(2).child(0).text();
-        String usdInitialApCost = usdJasperReportSummary.getReportHtmlPart().getElementsByAttributeValue("colspan", "2").get(20).child(0).text();
-
-        assertThat(gbpCurrencySettingAboveChart, is(not(equalTo(usdCurrencySettingAboveChart))));
-        assertThat(gbpInitialApCost, is(not(equalTo(usdInitialApCost))));
+        jasperApiUtils.spendAnalysisReportGenericCurrencyTest(Arrays.asList(7, 2, 2, 19));
     }
 }
